refactor(cast): finish command cleanup
Borrow concatenation input and log query state, keep helpers internal,
and finish grouping CLI coverage. Add regression cases for conversion
boundaries, mapping slots, and source output.
