+++ .changelog/cast-command-implementation-cleanup.md
@@ -1,0 +1,5 @@
+---
+cast: patch
+---
+
+Simplified Cast command implementations and reorganized CLI regression coverage.
